WARNING: work in progress ... new filesystems API with caching support and important fixes (details bellow): works on linux, breaks other platforms (will be fixed one by one)
New: Linux: cache per-filesystem mount table status and refresh only if changed (some filesystem mounted/unmounted)

New: It is now possible to use NFS/CIFS/SSHFS/etc. connection strings for filesystem path. Examples:
        check filesystem mynfs with path "192.168.1.123:/home/myuser"
        check filesystem mycifs with path "//192.168.1.123/documents"

Fixed: Previous Monit versions may be unable to detect that the filesystem is not
mounted in the case that mountpoint was used as a filesystem path. The reason is,
that the mountpoint itself resides on some filesystem (for example root filesystem)
and thus even if the filesystem is not mounted, the directory may still exist - Monit
didn't require exact match in the mount table and allowed filesystem subdirectory or even
regular file to be used as filesystem path. The exact mountpoint match is now required,
so Monit can check if the filesystem is mounted. This fix doesn't affect the configuration
with block/character device.